handle invitation page

+++ graphql/types/Invitation.ts
@@ -1,3 +1,4 @@
+import { Invitation } from "@prisma/client";
 import { builder } from "../builder";
 
 builder.prismaObject("Invitation", {
@@ -10,6 +11,17 @@ builder.prismaObject("Invitation", {
       type: "DateTime",
     }),
     email: t.exposeString("email"),
+    name: t.string({
+      nullable: true,
+      resolve: async (parent, args, ctx, info) => {
+        const user = await prisma.user.findUnique({
+          where: {
+            email: parent.email,
+          },
+        });
+        return user?.name || null;
+      },
+    }),
     workspace: t.relation("workspace"),
     workspaceId: t.exposeID("workspaceId"),
   }),
@@ -42,7 +54,7 @@ builder.queryFields((t) => ({
     // },
     resolve: async (query, parent, args, ctx, info) => {
       // console.log(ctx.user?.id);
-      return prisma.invitation.findMany({
+      const invitation = await prisma.invitation.findMany({
         ...query,
         where: {
           workspace: {
@@ -50,6 +62,22 @@ builder.queryFields((t) => ({
           },
         },
       });
+
+      // add name of user with this email to invitation
+      const invitations = await Promise.all(
+        invitation.map(async (invitation) => {
+          const user = await prisma.user.findUnique({
+            where: {
+              email: invitation.email,
+            },
+          });
+          return {
+            ...invitation,
+            user: user,
+          };
+        })
+      );
+      return invitations;
     },
   }),
   recivedInvitations: t.prismaField({
@@ -62,6 +90,7 @@ builder.queryFields((t) => ({
 
     resolve: async (query, parent, args, ctx, info) => {
       // console.log(ctx.user?.email);
+
       return prisma.invitation.findMany({
         ...query,
         where: {
@@ -78,12 +107,54 @@ builder.mutationFields((t) => ({
     args: {
       email: t.arg.string({
         required: true,
+        validate: {
+          email: true,
+        },
       }),
       workspaceId: t.arg.id({
         required: true,
       }),
     },
     resolve: async (query, parent, args, ctx, info) => {
+      // check if exists user with this email
+      const user = await prisma.user.findUnique({
+        where: {
+          email: args.email,
+        },
+      });
+      if (!user) {
+        throw new Error("User with this email does not exist");
+      }
+      // check if user is already invited to this workspace
+      const invitation = await prisma.invitation.findFirst({
+        where: {
+          workspaceId: args.workspaceId.toString(),
+          email: args.email,
+        },
+      });
+      if (invitation) {
+        throw new Error("User is already invited to this workspace");
+      }
+      // check if user is already member of this workspace
+      const membership = await prisma.workspaceMembership.findFirst({
+        where: {
+          workspaceId: args.workspaceId.toString(),
+          userId: user.id,
+        },
+      });
+      if (membership) {
+        throw new Error("User is already member of this workspace");
+      }
+      // check if user is already owner of this workspace
+      const workspace = await prisma.workspace.findUnique({
+        where: {
+          id: args.workspaceId.toString(),
+        },
+      });
+      if (workspace?.ownerId === user.id) {
+        throw new Error("User is already owner of this workspace");
+      }
+
       return prisma.invitation.create({
         ...query,
         data: {
@@ -127,6 +198,27 @@ builder.mutationFields((t) => ({
       }),
     },
     resolve: async (query, parent, args, ctx, info) => {
+      // check if ctx.user.id === invitation.workspace.ownerId or ctx.user.email === invitation.email
+      const invitation = await prisma.invitation.findFirst({
+        ...query,
+        where: {
+          OR: [
+            {
+              workspace: {
+                ownerId: ctx.user?.id,
+              },
+            },
+            {
+              email: ctx.user?.email,
+            },
+          ],
+        },
+      });
+
+      if (!invitation) {
+        throw new Error("You are not allowed to delete this invitation");
+      }
+
       return prisma.invitation.delete({
         ...query,
         where: {
